refactor: move setup functions to a seperate module
